+++ opentelemetry-aws/src/detector/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "detector-aws-lambda")]
+mod lambda;
+#[cfg(feature = "detector-aws-lambda")]
+pub use lambda::LambdaResourceDetector;